pyproject.toml
[tool.poetry]
name = "yasno-for-qa-guru"
version = "0.1.0"
description = ""
authors = ["Daria Chaenkova <dash.kustowa2010@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.6"
pytest = "^8.3.2"
selene = ">=2.0.0rc9"
allure-pytest = "^2.13.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
